refactor: search dialog 내부 로직 개별 훅으로 분리
