+++ src/components/SearchDialog/index.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useRef } from "react";
 import {
   Box,
   HStack,
@@ -16,20 +16,20 @@ import {
   Tr,
   VStack,
 } from "@chakra-ui/react";
-import { Lecture, ProcessedLecture, SearchOption } from "../../types";
-import { parseSchedule } from "../../lib/utils";
-import { PAGE_SIZE } from "../../constants";
-import { useAutoCallback } from "../../hooks/useAutoCallback.ts";
-import { fetchAllLectures } from "../../lib/api.ts";
-import { TimeFilter } from "./TimeFilter.tsx";
-import { MajorFilter } from "./MajorFilter.tsx";
-import { SearchItem } from "./SearchItem.tsx";
 import { useScheduleStore } from "../../store/scheduleStore";
-import { useDebounce } from "../../hooks/useDebounce.ts";
+import { useAutoCallback } from "../../hooks/useAutoCallback.ts";
+import { useLectureData } from "../../hooks/useLectureData.ts";
+import { useLectureFilter } from "../../hooks/useLectureFilter.ts";
+import { useInfiniteScroll } from "../../hooks/useInfiniteScroll.ts";
+import { Lecture } from "../../types";
+import { parseSchedule } from "../../lib/utils";
 import { GradeFilter } from "./GradeFilter.tsx";
 import { DayFilter } from "./DayFilter.tsx";
 import { SearchFilter } from "./SearchFilter.tsx";
 import { CreditFilter } from "./CreditFilter.tsx";
+import { TimeFilter } from "./TimeFilter.tsx";
+import { MajorFilter } from "./MajorFilter.tsx";
+import { SearchItem } from "./SearchItem.tsx";
 
 interface Props {
   searchInfo: {
@@ -41,83 +41,14 @@ interface Props {
 }
 
 const SearchDialog = ({ searchInfo, onClose }: Props) => {
+  const { processedLectures, allMajors } = useLectureData();
+  const { searchOptions, changeSearchOption, filteredLectures } =
+    useLectureFilter(processedLectures);
+  const { visibleLectures, loaderRef, hasMore } = useInfiniteScroll(filteredLectures);
+
   const setSchedulesMap = useScheduleStore((state) => state.setSchedulesMap);
 
-  const observerRef = useRef<IntersectionObserver>(null);
   const loaderWrapperRef = useRef<HTMLDivElement>(null);
-
-  const [lectures, setLectures] = useState<Lecture[]>([]);
-  const [processedLectures, setProcessedLectures] = useState<ProcessedLecture[]>([]);
-  const [page, setPage] = useState(1);
-  const [searchOptions, setSearchOptions] = useState<SearchOption>({
-    query: "",
-    grades: [],
-    days: [],
-    times: [],
-    majors: [],
-  });
-
-  const debouncedSearchOptions = useDebounce(searchOptions, 300);
-
-  useEffect(() => {
-    if (lectures && lectures.length > 0) {
-      const parsedData = lectures.map((lecture) => ({
-        ...lecture,
-        parsedSchedule: lecture.schedule ? parseSchedule(lecture.schedule) : [],
-      }));
-      setProcessedLectures(parsedData);
-    }
-  }, [lectures]);
-
-  const filteredLectures = useMemo(() => {
-    const { query = "", credits, grades, days, times, majors } = debouncedSearchOptions;
-    return processedLectures
-      .filter(
-        (lecture) =>
-          lecture.title.toLowerCase().includes(query.toLowerCase()) ||
-          lecture.id.toLowerCase().includes(query.toLowerCase())
-      )
-      .filter((lecture) => grades.length === 0 || grades.includes(lecture.grade))
-      .filter((lecture) => majors.length === 0 || majors.includes(lecture.major))
-      .filter((lecture) => !credits || lecture.credits.startsWith(String(credits)))
-      .filter((lecture) => {
-        if (days.length === 0) return true;
-        return lecture.parsedSchedule.some((s) => days.includes(s.day));
-      })
-      .filter((lecture) => {
-        if (times.length === 0) return true;
-        return lecture.parsedSchedule.some((s) => s.range.some((t) => times.includes(t)));
-      });
-  }, [processedLectures, debouncedSearchOptions]);
-
-  const lastPage = Math.ceil(filteredLectures.length / PAGE_SIZE);
-  const visibleLectures = filteredLectures.slice(0, page * PAGE_SIZE);
-  const allMajors = useMemo(() => {
-    return [...new Set(lectures.map((lecture) => lecture.major))];
-  }, [lectures]);
-
-  const hasMore = page < lastPage;
-
-  const changeSearchOption = useCallback(
-    (field: keyof SearchOption, value: any) => {
-      setPage(1);
-      setSearchOptions((prevOptions) => ({ ...prevOptions, [field]: value }));
-    },
-    [setSearchOptions]
-  );
-
-  const loaderRef = useCallback(
-    (node: HTMLDivElement) => {
-      const observer = observerRef.current;
-      if (observer) {
-        observer.disconnect();
-        if (node) {
-          observer.observe(node);
-        }
-      }
-    },
-    [observerRef.current]
-  );
 
   const addSchedule = useAutoCallback((lecture: Lecture) => {
     if (!searchInfo) return;
@@ -132,53 +63,22 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
     setSchedulesMap((prev: any) => {
       return {
         ...prev,
-        [tableId]: [...prev[tableId], ...schedules],
+        [tableId]: [...(prev[tableId] || []), ...schedules],
       };
     });
-
     onClose();
   });
 
   useEffect(() => {
-    fetchAllLectures().then((results) => {
-      setLectures(results.flatMap((result) => result.data));
-    });
-  }, []);
-
-  const handleLoadMore = useCallback(() => {
-    if (hasMore) {
-      setPage((prevPage) => prevPage + 1);
-    }
-  }, [hasMore]);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) {
-          handleLoadMore();
-        }
-      },
-      { threshold: 0.1 }
-    );
-    observerRef.current = observer;
-  }, [handleLoadMore]);
-
-  useEffect(() => {
     if (searchInfo) {
-      setSearchOptions((prev) => ({
-        ...prev,
-        days: searchInfo?.day ? [searchInfo.day] : [],
-        times: searchInfo?.time ? [searchInfo.time] : [],
-      }));
-      setPage(1);
+      changeSearchOption("days", searchInfo?.day ? [searchInfo.day] : []);
+      changeSearchOption("times", searchInfo?.time ? [searchInfo.time] : []);
     }
-  }, [searchInfo]);
+  }, [searchInfo, changeSearchOption]);
 
   useEffect(() => {
-    if (page === 1) {
-      loaderWrapperRef.current?.scrollTo(0, 0);
-    }
-  }, [page, filteredLectures.length]);
+    loaderWrapperRef.current?.scrollTo(0, 0);
+  }, [filteredLectures]);
 
   return (
     <Modal isOpen={Boolean(searchInfo)} onClose={onClose} size="6xl">
